Hand size and size of hand
Added larger card png for when it's in the hand so we can have more distinct text on it.
Set the panel in card_frame.tscn to be clear so we don't see it stretching in the hand.

Added abstract class intended to lay out what should be in a shop button script
Added shop button for extending max hand by 1